core: refactor deployment subgenerators
