+++ generators/kubernetes/files.js
@@ -80,7 +80,7 @@ function writeFiles() {
                 this.template('console/jhipster-logstash.yml.ejs', 'console/jhipster-logstash.yml');
                 this.template('console/jhipster-console.yml.ejs', 'console/jhipster-console.yml');
                 this.template('console/jhipster-dashboard-console.yml.ejs', 'console/jhipster-dashboard-console.yml');
-                if (this.composeApplicationType === 'microservice') {
+                if (this.deploymentApplicationType === 'microservice') {
                     this.template('console/jhipster-zipkin.yml.ejs', 'console/jhipster-zipkin.yml');
                 }
             }
