+++ generators/openshift/templates/deployment.yml.ejs
@@ -243,7 +243,7 @@ objects:
               value: "true"
             - name: JHIPSTER_LOGGING_LOGSTASH_HOST
               value: ${JHIPSTER_LOGSTASH}
-              <%_ if (composeApplicationType === 'microservice') { _%>
+              <%_ if (deploymentApplicationType === 'microservice') { _%>
             - name: SPRING_ZIPKIN_ENABLED
               value: "true"
             - name: SPRING_ZIPKIN_BASE_URL
