+++ generators/rancher-compose/templates/docker-compose.yml.ejs
@@ -131,7 +131,7 @@ services:
         image: <%= DOCKER_JHIPSTER_CONSOLE %>
         ports:
             - 5601:5601
-    <%_ if (composeApplicationType === 'microservice') { _%>
+    <%_ if (deploymentApplicationType === 'microservice') { _%>
     # Uncomment this section to enable Zipkin
     #jhipster-zipkin:
     #    image: <%= DOCKER_JHIPSTER_ZIPKIN %>
