+++ generators/rancher-compose/templates/rancher-compose.yml.ejs
@@ -60,7 +60,7 @@ services:
         scale: 1
     jhipster-console:
         scale: 1
-    <%_ if (composeApplicationType === 'microservice') { _%>
+    <%_ if (deploymentApplicationType === 'microservice') { _%>
     # Uncomment this section to enable Zipkin
     #jhipster-zipkin:
     #    scale: 1
