+++ generators/utils.js
@@ -23,6 +23,7 @@ const ejs = require('ejs');
 const _ = require('lodash');
 const jhiCore = require('jhipster-core');
 const fs = require('fs');
+const crypto = require('crypto');
 
 const constants = require('./generator-constants');
 
@@ -42,7 +43,9 @@ module.exports = {
     copyObjectProps,
     decodeBase64,
     getAllJhipsterConfig,
-    getDBTypeFromDBValue
+    getDBTypeFromDBValue,
+    getBase64Secret,
+    getRandomHex
 };
 
 /**
@@ -442,4 +445,20 @@ function getDBTypeFromDBValue(db) {
         return 'sql';
     }
     return db;
+}
+
+/**
+ * Get a random hex string
+ * @param {int} len the length to use, defaults to 50
+ */
+function getRandomHex(len = 50) {
+    return crypto.randomBytes(len).toString('hex');
+}
+/**
+ * Generates a base64 secret from given string or random hex
+ * @param {string} value the value used to get base64 secret
+ * @param {int} len the length to use for random hex, defaults to 50
+ */
+function getBase64Secret(value, len = 50) {
+    return Buffer.from(value || getRandomHex(len)).toString('base64');
 }
